+++ honestcash-v2/src/app/modules/editor/editor-core/editor-core.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { CoreModule } from '../../../core';
 import { EditorCoreComponent } from './editor-core.component';
-import { PostService } from '../../../services/post.service';
+import { PostService } from '../../../core/services/post.service';
 import { EditorService } from './editor.service';
 
 describe('EditorCoreComponent', () => {
