+++ honestcash-v2/src/app/store/wallet/wallet.effects.ts
@@ -7,7 +7,7 @@ import {
     WalletActionTypes,
     WalletSetup,
 } from './wallet.actions';
-import { WalletService } from 'app/services/wallet.service';
+import { WalletService } from '../../core/services/wallet.service';
 
 @Injectable()
 export class WalletEffects {
